Added Survey Start Page and Tests
Added the start page for the survey. This passes the user type and user name to the other controllers to be submitted at the end of the survey. It will check if the user name has been entered and will not allow the user to take the survey without a name